Move SwitchPuzzleDialog to new file
